requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.2
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.21
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==41.0.7
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.8.0
fastapi==0.108.0
filelock==3.13.1
firebase-admin==6.3.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
google-analytics-admin==0.22.1
google-analytics-data==0.18.2
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
httpx-sse==0.4.0
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
json-spec==0.11.0
jsonpatch==1.33
jsonpointer==2.4
jsontool==0.2.1
kubernetes==28.1.0
langchain==0.0.354
langchain-cli==0.0.20
langchain-community==0.0.8
langchain-core==0.1.6
langserve==0.0.37
langsmith==0.0.77
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.6.1
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.9.10
overrides==7.4.0
packaging==23.2
posthog==3.1.0
proto-plus==1.23.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pulsar-client==3.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.24
sse-starlette==1.8.2
starlette==0.32.0.post1
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.0
tomlkit==0.12.3
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
